Added API tests
