+++ api/__init__.py
@@ -7,7 +7,9 @@ from flask import Flask
 from flask_restful import Api
 
 app = Flask(__name__)
-api = Api(app)
+app.url_map.strict_slashes = False
+
+api = Api(app, catch_all_404s=True)
 
 # # Recursive Imports
 from api import errors, calls
