+++ requirements.txt
@@ -1,6 +1,8 @@
 aniso8601==8.0.0
 astroid==2.3.0
 bandit==1.6.2
+certifi==2019.9.11
+chardet==3.0.4
 Click==7.0
 colorama==0.4.1
 entrypoints==0.3
@@ -9,6 +11,7 @@ Flask==1.1.1
 Flask-RESTful==0.3.7
 gitdb2==2.0.5
 GitPython==3.0.2
+idna==2.8
 isort==4.3.21
 itsdangerous==1.1.0
 Jinja2==2.10.1
@@ -23,9 +26,11 @@ pylint==2.4.1
 python-dotenv==0.10.3
 pytz==2019.2
 PyYAML==5.1.2
+requests==2.22.0
 six==1.12.0
 smmap2==2.0.5
 stevedore==1.31.0
 typed-ast==1.4.0
+urllib3==1.25.6
 Werkzeug==0.16.0
 wrapt==1.11.2
