+++ tests/README.md
@@ -1,2 +1,5 @@
 You are able to call all tests on the command line at project root using:
-    $ python -m unittest tests
+
+```bash 
+$ python -m unittest tests
+```
