+++ tests/__init__.py
@@ -7,3 +7,5 @@ Call all tests on the command line at project root using:
 # # Internal Modules
 from .test_to_cardinal import TestToCardinalNumber
 from .test_to_cardinal import TestToCardinalTrio
+from .test_api import TestApiCalls
+# from .test_api import TestApiErrors
